+++ erppeek.py
@@ -59,8 +59,8 @@ except ImportError:
         return _convert(node_or_string)
 
 
-__version__ = '0.8'
-__all__ = ['Client', 'read_config']
+__version__ = '0.9.dev0'
+__all__ = ['Client', 'Model', 'Record', 'RecordList', 'Service', 'read_config']
 
 CONF_FILE = 'erppeek.ini'
 DEFAULT_URL = 'http://localhost:8069'
@@ -80,7 +80,8 @@ Usage (main commands):
     read(obj, domain, fields=None, offset=0, limit=None, order=None)
                                     # Return values for the fields
 
-    model(name)                     # List models matching pattern
+    models(name)                    # List models matching pattern
+    model(name)                     # Return a Model instance
     keys(obj)                       # List field names of the model
     fields(obj, names=None)         # Return details for the fields
     field(obj, name)                # Return details for the field
@@ -139,7 +140,44 @@ _methods_6_1 = {
 #  - 'common': ['get_available_updates', 'get_migration_scripts', 'set_loglevel']
 
 
+def mixedcase(s, _cache={}):
+    """Convert to MixedCase.
+
+    >>> mixedcase('res.company')
+    'ResCompany'
+    """
+    try:
+        return _cache[s]
+    except KeyError:
+        _cache[s] = s = ''.join([w.capitalize() for w in s.split('.')])
+    return s
+
+
+def lowercase(s, _sub=re.compile('[A-Z]').sub,
+              _repl=(lambda m: '.' + m.group(0).lower()), _cache={}):
+    """Convert to lowercase with dots.
+
+    >>> lowercase('ResCompany')
+    'res.company'
+    """
+    try:
+        return _cache[s]
+    except KeyError:
+        _cache[s] = s = _sub(_repl, s).lstrip('.')
+        return s
+
+
 def read_config(section=None):
+    """Read the environment settings from the configuration file.
+
+    The config file ``erppeek.ini`` contains a `section` for each environment.
+    Each section provides parameters for the connection: ``host``, ``port``,
+    ``database``, ``user`` and (optional) ``password``.  Default values are
+    read from the ``[DEFAULT]`` section.  If the ``password`` is not in the
+    configuration file, it is requested on login.
+    Return a tuple ``(server, db, user, password or None)``.
+    Without argument, it returns the list of configured environments.
+    """
     p = configparser.SafeConfigParser()
     with open(Client._config_file) as f:
         p.readfp(f)
@@ -159,11 +197,11 @@ def issearchdomain(arg):
     """Check if the argument is a search domain.
 
     Examples:
-      - [('name', '=', 'mushroom'), ('state', '!=', 'draft')]
-      - ['name = mushroom', 'state != draft']
-      - []
-      - 'state != draft'
-      - ('state', '!=', 'draft')
+      - ``[('name', '=', 'mushroom'), ('state', '!=', 'draft')]``
+      - ``['name = mushroom', 'state != draft']``
+      - ``[]``
+      - ``'state != draft'``
+      - ``('state', '!=', 'draft')``
     """
     return isinstance(arg, (list, tuple, basestring)) and not (arg and (
             # Not a list of ids: [1, 2, 3]
@@ -206,6 +244,15 @@ def searchargs(params, kwargs=None, context=None):
 
 
 class Service(ServerProxy):
+    """A wrapper around XML-RPC endpoints.
+
+    The connected endpoints are exposed on the Client instance.
+    The `server` argument is the URL of the server (scheme+host+port).
+    The `endpoint` argument is the last part of the URL
+    (examples: ``"object"``, ``"db"``).  The `methods` is the list of methods
+    which should be exposed on this endpoint.  Use ``dir(...)`` on the
+    instance to list them.
+    """
     def __init__(self, server, endpoint, methods):
         uri = server + '/xmlrpc/' + endpoint
         ServerProxy.__init__(self, uri, allow_none=True)
@@ -228,6 +275,14 @@ class Service(ServerProxy):
 
 
 class Client(object):
+    """Connection to an OpenERP instance.
+
+    This is the top level object.
+    The `server` is the URL of the instance, like ``http://localhost:8069``.
+    The `db` is the name of the database and the `user` should exist in the
+    table ``res.users``.  If the `password` is not provided, it will be
+    asked on login.
+    """
     _config_file = os.path.join(os.path.curdir, CONF_FILE)
 
     def __init__(self, server, db, user, password=None):
@@ -255,9 +310,16 @@ class Client(object):
         self._report = get_proxy('report')
         # Try to login
         self._login(user, password)
+        self._models = {}
 
     @classmethod
     def from_config(cls, environment):
+        """Create a connection to a defined environment.
+
+        Read the settings from the section ``[environment]`` in the
+        ``erppeek.ini`` file and return a connected :class:`Client`.
+        See :func:`read_config` for details of the configuration file format.
+        """
         client = cls(*read_config(environment))
         client._environment = environment
         return client
@@ -266,6 +328,10 @@ class Client(object):
         return "<Client '%s#%s'>" % (self._server, self._db)
 
     def login(self, user, password=None):
+        """Switch `user`.
+
+        If the `password` is not provided, it will be asked.
+        """
         (uid, password) = self._auth(user, password)
         if uid is False:
             print('Error: Invalid username or password')
@@ -374,6 +440,13 @@ class Client(object):
         cls.connect = connect
 
     def execute(self, obj, method, *params, **kwargs):
+        """Wrapper around ``object.execute`` RPC method.
+
+        Argument `method` is the name of an ``osv.osv`` method or
+        a method available on this `obj`.
+        Method `params` are allowed.  If needed, keyword
+        arguments are collected in `kwargs`.
+        """
         assert isinstance(obj, basestring) and isinstance(method, basestring)
         context = kwargs.pop('context', None)
         if method in ('read', 'name_get'):
@@ -404,10 +477,25 @@ class Client(object):
         return self._execute(obj, method, *params)
 
     def exec_workflow(self, obj, signal, obj_id):
+        """Wrapper around ``object.exec_workflow`` RPC method.
+
+        Argument `obj` is the name of the model.  The `signal`
+        is sent to the object identified by its integer ``id`` `obj_id`.
+        """
         assert isinstance(obj, basestring) and isinstance(signal, basestring)
         return self._exec_workflow(obj, signal, obj_id)
 
     def wizard(self, name, datas=None, action='init', context=None):
+        """Wrapper around ``wizard.create`` and ``wizard.execute``
+        RPC methods.
+
+        If only `name` is provided, a new wizard is created and its ``id`` is
+        returned.  If `action` is not ``"init"``, then the action is executed.
+        In this case the `name` is either an ``id`` or a string.
+        If the `name` is a string, the wizard is created before the execution.
+        The optional `datas` argument provides data for the action.
+        The optional `context` argument is passed to the RPC method.
+        """
         if isinstance(name, int_types):
             wiz_id = name
         else:
@@ -438,6 +526,7 @@ class Client(object):
         for mod in mods:
             print('  %(state)s\t%(name)s' % mod)
 
+        self._models.clear()
         if self.major_version == '5.0':
             # Wizard "Apply Scheduled Upgrades"
             rv = self.wizard('module.upgrade', action='start')
@@ -448,37 +537,40 @@ class Client(object):
             self.execute('base.module.upgrade', 'upgrade_module', [])
 
     def upgrade(self, *modules):
-        # Button "Schedule Upgrade"
+        """Press the button ``Upgrade``."""
         return self._upgrade(modules, button='button_upgrade')
 
     def install(self, *modules):
-        # Button "Schedule for Installation"
+        """Press the button ``Install``."""
         return self._upgrade(modules, button='button_install')
 
     def uninstall(self, *modules):
-        # Button "Uninstall (beta)"
+        """Press the button ``Uninstall``."""
         return self._upgrade(modules, button='button_uninstall')
 
     def search(self, obj, *params, **kwargs):
+        """Filter the records in the `domain`, return the ``ids``."""
         return self.execute(obj, 'search', *params, **kwargs)
 
     def count(self, obj, domain=None):
+        """Count the records in the `domain`."""
         return self.execute(obj, 'search_count', domain or [])
 
     def read(self, obj, *params, **kwargs):
-        """Wrapper for client.execute(obj, 'read', [...], ('a', 'b')).
+        """Wrapper for ``client.execute(obj, 'read', [...], ('a', 'b'))``.
 
-        The first argument is the 'model' name (example: 'res.partner')
+        The first argument `obj` is the model name (example: ``"res.partner"``)
 
-        The second argument, 'domain', accepts:
-         - [('name', '=', 'mushroom'), ('state', '!=', 'draft')]
-         - ['name = mushroom', 'state != draft']
-         - []
+        The second argument, `domain`, accepts:
+         - ``[('name', '=', 'mushroom'), ('state', '!=', 'draft')]``
+         - ``['name = mushroom', 'state != draft']``
+         - ``[]``
+         - or a list of ids ``[1, 2, 3]``
 
-        The third argument, 'fields', accepts:
-         - ('street', 'city')
-         - 'street city'
-         - '%(street)s %(city)s'
+        The third argument, `fields`, accepts:
+         - ``('street', 'city')``
+         - ``'street city'``
+         - ``'%(street)s %(city)s'``
         """
         fmt = None
         if len(params) > 1 and isinstance(params[1], basestring):
@@ -504,17 +596,66 @@ class Client(object):
             return res[fields[0]]
         return res
 
-    def model(self, name):
+    def _model(self, name):
+        try:
+            return self._models[name]
+        except KeyError:
+            # m = Model(self, name)
+            m = object.__new__(Model)
+        m._init(self, name)
+        self._models[name] = m
+        return m
+
+    def models(self, name=''):
+        """Return a dictionary of models.
+
+        The argument `name` is a pattern to filter the models returned.
+        If omitted, all models are returned.
+        Keys are camel case names of the models.
+        Values are instances of :class:`Model`.
+
+        The return value can be used to declare the models in the global
+        namespace:
+
+        >>> globals().update(client.models('res.'))
+        """
         domain = [('model', 'like', name)]
         models = self.execute('ir.model', 'read', domain, ('model',))
+        names = [m['model'] for m in models]
+        return dict([(mixedcase(name), self._model(name)) for name in names])
+
+    def model(self, name):
+        """Return a :class:Model instance.
+
+        The argument `name` is the name of the model.
+        """
+        try:
+            return self._models[name]
+        except KeyError:
+            models = self.models(name)
+        if name in self._models:
+            return self._models[name]
         if models:
-            return sorted([m['model'] for m in models])
+            errmsg = 'Model not found.  These models exist:'
+        else:
+            errmsg = 'Model not found: %s' % (name,)
+        raise RuntimeError('\n * '.join([errmsg] + models.values()))
 
     def modules(self, name='', installed=None):
+        """Return a dictionary of modules.
+
+        The optional argument `name` is a pattern to filter the modules.
+        If the boolean argument `installed` is :const:`True`, the modules
+        which are "Not Installed" or "Not Installable" are omitted.  If
+        the argument is :const:`False`, only these modules are returned.
+        If argument `installed` is omitted, all modules are returned.
+        The return value is a dictionary where module names are grouped in
+        lists according to their ``state``.
+        """
         domain = [('name', 'like', name)]
         if installed is not None:
-            op = installed and '!=' or '='
-            domain.append(('state', op, 'uninstalled'))
+            op = installed and 'not in' or 'in'
+            domain.append(('state', op, ['uninstalled', 'uninstallable']))
         mods = self.read('ir.module.module', domain, 'name state')
         if mods:
             res = {}
@@ -526,25 +667,28 @@ class Client(object):
             return res
 
     def keys(self, obj):
-        obj_keys = self.execute(obj, 'fields_get_keys')
-        return obj_keys and sorted(obj_keys)
+        """Wrapper for :meth:`Model.keys` method."""
+        return self.model(obj).keys()
 
     def fields(self, obj, names=None):
-        return self.execute(obj, 'fields_get', names)
+        """Wrapper for :meth:`Model.fields` method."""
+        return self.model(obj).fields(names=names)
 
     def field(self, obj, name):
-        return self.execute(obj, 'fields_get', (name,))[name]
+        """Wrapper for :meth:`Model.fields` method."""
+        return self.model(obj).field(name)
 
     def access(self, obj, mode='read'):
-        try:
-            self._execute('ir.model.access', 'check', obj, mode)
-            return True
-        except (TypeError, Fault):
-            return False
+        """Wrapper for :meth:`Model.access` method."""
+        return self.model(obj).access(mode=mode)
 
     def __getattr__(self, method):
         if method.startswith('__'):
             raise AttributeError("'Client' object has no attribute %r" % method)
+        if not method.islower():
+            rv = self.model(lowercase(method))
+            self.__dict__[method] = rv
+            return rv
         # miscellaneous object methods
         def wrapper(self, obj, *params, **kwargs):
             """Wrapper for client.execute(obj, %r, *params, **kwargs)."""
@@ -552,6 +696,309 @@ class Client(object):
         wrapper.__name__ = method
         wrapper.__doc__ %= method
         return wrapper.__get__(self, type(self))
+
+
+class Model(object):
+    """The class for OpenERP models."""
+
+    def __new__(cls, client, name):
+        return client.model(name)
+
+    def _init(self, client, name):
+        self.client = client
+        self._name = name
+        self._execute = functools.partial(client.execute, name)
+        self.search = functools.partial(client.search, name)
+        self.count = functools.partial(client.count, name)
+        self.read = functools.partial(client.read, name)
+
+    def __repr__(self):
+        return "<Model '%s'>" % (self._name,)
+
+    @property
+    def _keys(self):
+        obj_keys = self._execute('fields_get_keys')
+        obj_keys.sort()
+        self.__dict__['_keys'] = obj_keys
+        return obj_keys
+
+    @property
+    def _fields(self):
+        fields = self._execute('fields_get')
+        self.__dict__['_fields'] = fields
+        return fields
+
+    def keys(self):
+        """Return the keys of the model."""
+        return self._keys
+
+    def fields(self, names=None):
+        """Return a dictionary of the fields of the model.
+
+        Optional argument `names` is a sequence of field names or
+        a space separated string of these names.
+        If omitted, all fields are returned.
+        """
+        if names is None:
+            return self._fields
+        if isinstance(names, basestring):
+            names = names.split()
+        return dict([(k, v) for (k, v) in self._fields.items() if k in names])
+
+    def field(self, name):
+        """Return the field properties for field `name`."""
+        return self._fields[name]
+
+    def access(self, mode="read"):
+        """Check if the user has access to this model.
+
+        Optional argument `mode` is the access mode to check.  Valid values
+        are ``read``, ``write``, ``create`` and ``unlink``. If omitted,
+        the ``read`` mode is checked.  Return a boolean.
+        """
+        try:
+            self.client._execute('ir.model.access', 'check', self._name, mode)
+            return True
+        except (TypeError, Fault):
+            return False
+
+    def browse(self, ids, context=None):
+        """Return a :class:`Record` or a :class:`RecordList`.
+
+        The argument `ids` accepts a single integer ``id``, a list of ids
+        or a search domain.
+        If it is a single integer, the return value is a :class:`Record`.
+        Otherwise, the return value is a :class:`RecordList`.
+        """
+        if isinstance(ids, int_types):
+            return Record(self, ids, context=context)
+        if issearchdomain(ids):
+            ids = self._execute('search', ids, context=context)
+        return RecordList(self, ids, context=context)
+    # alias
+    get = browse
+
+    def create(self, values, context=None):
+        """Create a :class:`Record`.
+
+        The argument `values` is a dictionary of values which are used to
+        create the record.  The newly created :class:`Record` is returned.
+        """
+        new_id = self._execute('create', values, context=context)
+        return Record(self, new_id, context=context)
+
+    def __getattr__(self, method):
+        if method in ('_keys', '_fields') or method.startswith('__'):
+            raise AttributeError("'Model' object has no attribute %r" % method)
+        def wrapper(self, *params, **kwargs):
+            """Wrapper for client.execute(%r, %r, *params, **kwargs)."""
+            return self._execute(method, *params, **kwargs)
+        wrapper.__name__ = method
+        wrapper.__doc__ %= (self._name, method)
+        self.__dict__[method] = mobj = wrapper.__get__(self, type(self))
+        return mobj
+
+
+class RecordList(object):
+    """A sequence of OpenERP :class:`Record`."""
+
+    def __init__(self, res_model, ids, context=None):
+        self._model = res_model
+        self._ids = ids
+        self._context = context
+
+    def __repr__(self):
+        if len(self._ids) > 16:
+            ids = 'length=%d' % len(self._ids)
+        else:
+            ids = self._ids
+        return "<RecordList '%s,%s'>" % (self._model._name, ids)
+
+    def __dir__(self):
+        return ['__getitem__', 'read', 'write', 'unlink',
+                '_context', '_ids', '_model']
+
+    def __getitem__(self, key):
+        if isinstance(key, slice):
+            return RecordList(
+                self._model, self._ids[key], context=self._context)
+        return Record(self._model, self._ids[key], context=self._context)
+
+    def __getattr__(self, attr):
+        if attr.startswith('__'):
+            errmsg = "'RecordList' object has ""no attribute %r" % attr
+            raise AttributeError(errmsg)
+        model_name = self._model._name
+        context = self._context
+        execute = self._model.client.execute
+        def wrapper(self, *params, **kwargs):
+            """Wrapper for client.execute(%r, %r, [...], *params, **kwargs)."""
+            if context:
+                kwargs.setdefault('context', context)
+            return execute(model_name, attr, self._ids, *params, **kwargs)
+        wrapper.__name__ = attr
+        wrapper.__doc__ %= (model_name, attr)
+        self.__dict__[attr] = mobj = wrapper.__get__(self, type(self))
+        return mobj
+
+
+class Record(object):
+    """A class for all OpenERP records.
+
+    It maps any OpenERP object.
+    The fields can be accessed through attributes.  The changes are immediately
+    saved in the database.
+    The ``many2one``, ``one2many`` and ``many2many`` attributes are followed
+    when the record is read.  Howeverm, when writing on these special fields,
+    use the appropriate syntax described in the official OpenERP documentation.
+    The attributes are evaluated lazily, and they are cached in the record.
+    The cache is invalidated if the :meth:`~Record.write` or the
+    :meth:`~Record.unlink` method is called.
+    """
+    def __init__(self, res_model, res_id, context=None):
+        self.__dict__.update({
+            'client': res_model.client,
+            '_model_name': res_model._name,
+            '_model': res_model,
+            '_context': context,
+            'id': res_id,
+        })
+
+    def __repr__(self):
+        return "<Record '%s,%d'>" % (self._model_name, self.id)
+
+    def __str__(self):
+        return self._name
+
+    @property
+    def _name(self):
+        try:
+            (id_name,) = self._model._execute('name_get', [self.id])
+            name = '[%d] %s' % (self.id, id_name[1])
+        except Exception:
+            name = '[%d] -' % (self.id,)
+        self.__dict__['_name'] = name
+        return name
+
+    @property
+    def _keys(self):
+        return self._model._keys
+
+    @property
+    def _fields(self):
+        return self._model._fields
+
+    def _clear_cache(self):
+        for key in self._model._keys:
+            if key != 'id' and key in self.__dict__:
+                delattr(self, key)
+
+    def read(self, fields=None, context=None):
+        """Read the `fields` of the :class:`Record`.
+
+        The argument `fields`, accepts different type of values:
+         - ``('street', 'city')``
+         - ``'street city'``
+         - ``'%(street)s %(city)s'``
+
+        If omitted, all fields are read.
+
+        Return a dictionary of values.
+        """
+        if context is None and self._context:
+            context = self._context
+        rv = self.client.read(self._model_name, self.id,
+                              fields, context=context)
+        self.__dict__.update(rv)
+        return rv
+
+    def write(self, values, context=None):
+        """Write the `values` in the :class:`Record`."""
+        if context is None and self._context:
+            context = self._context
+        rv = self.client.write(self._model_name, 'write', [self.id],
+                               values, context=context)
+        self._clear_cache()
+        return rv
+
+    def unlink(self, context=None):
+        """Delete the current :class:`Record` from the database."""
+        if context is None and self._context:
+            context = self._context
+        rv = self.client.unlink(self._model_name, 'unlink', [self.id],
+                                context=context)
+        self._clear_cache()
+        return rv
+
+    def copy(self, default=None, context=None):
+        """Copy a record and return the new :class:`Record`.
+
+        The optional argument `default` is a mapping which overrides some
+        values of the new record.
+        """
+        if context is None and self._context:
+            context = self._context
+        new_id = self.client.copy(self._model_name, 'copy', self.id,
+                                  default=default, context=context)
+        return Record(self._model, new_id)
+
+    def __dir__(self):
+        return ['client', 'read', 'write', 'copy', 'unlink',
+                '_context', 'id', '_model', '_model_name',
+                '_name', '_keys', '_fields'] + self._model._keys
+
+    def __getitem__(self, attr):
+        try:
+            return self.__dict__[attr]
+        except KeyError:
+            if attr not in self._model._keys:
+                raise
+        value = self.client.read(self._model_name, self.id, attr)
+        rv = self._update({attr: value})
+        return rv[attr]
+
+    def _update(self, values):
+        new_values = {}
+        for key, value in values.items():
+            field = self._model._fields[key]
+            field_type = field['type']
+            if not value:
+                pass
+            elif field_type == 'many2one':
+                value = self.client.model(field['relation']).browse(
+                    value[0], context=self._context)
+            elif field_type in ('one2many', 'many2many'):
+                value = self.client.model(field['relation']).browse(
+                    value, context=self._context)
+            new_values[key] = value
+        self.__dict__.update(new_values)
+        return new_values
+
+    def __getattr__(self, attr):
+        if attr in self._model._keys:
+            return self[attr]
+        if attr.startswith('__'):
+            raise AttributeError("'Record' object has no attribute %r" % attr)
+        context = self._context
+        def wrapper(self, *params, **kwargs):
+            """Wrapper for client.execute(%r, %r, %d, *params, **kwargs)."""
+            if context:
+                kwargs.setdefault('context', context)
+            return self.client.execute(
+                self._model_name, attr, self.id, *params, **kwargs)
+        wrapper.__name__ = attr
+        wrapper.__doc__ %= (self._model_name, attr, self.id)
+        self.__dict__[attr] = mobj = wrapper.__get__(self, type(self))
+        return mobj
+
+    def __setattr__(self, attr, value):
+        if attr in self._model._keys:
+            assert attr != 'id'
+            self.write({attr: value})
+            if attr in self.__dict__:
+                delattr(self, attr)
+        else:
+            object.__setattr__(self, attr, value)
 
 
 def _interact(use_pprint=True, usage=USAGE):